Check for errors in BN_bn2dec()
If an oversize BIGNUM is presented to BN_bn2dec() it can cause
BN_div_word() to fail and not reduce the value of 't' resulting
in OOB writes to the bn_data buffer and eventually crashing.

Fix by checking return value of BN_div_word() and checking writes
don't overflow buffer.

Thanks to Shi Lei for reporting this bug.

CVE-2016-2182

Change-Id: I4f22198134c27c29511b1d6c7d09686b43cf2bd9
Reviewed-by: Tim Hudson <tjh@openssl.org>
Signed-off-by: spezi77 <spezi77@gmx.com>